requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5

SQLAlchemy==2.0.23
alembic==1.12.1

python-dotenv==1.0.1
requests==2.31.0
openai==1.3.7

PyPDF2==3.0.1
python-docx==1.1.0

werkzeug==2.3.7
